SafeCrash Class
The class that implements the majority of our error handling and streamlines it. It is a static method so to call it, simply write SafeCrash.shutDown(errorType, errorMessage, location); These are all taken as strings (description below) and the overall method handling is straightforward to implement. Simply make it the cause of action anytime an exception is caught. It's been working for me but little visual preferences can be fixed. Let Anjola know if you have any questions. 

        @param errorType: Type of error encountered. Ex. FileNotFoundException.
	@param errorMessage: Message for the user on the logistics of the error encountered. Ex. Unable to find file.
	@param location: location of the error in the format: Package, Class, Method.